requirements.txt
python-telegram-bot==22.0
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
pytz==2025.1 